cstdlib/cstring.h -> libk++/cstring.hpp
